Add portal thickness to improve accuracy

+++ source/geometry.hpp
@@ -6,8 +6,8 @@
 #include <iostream>
 #include <cmath>
 
-//using Scal = float;
-using Scal = double;
+using Scal = float;
+//using Scal = double;
 
 const Scal PI=atan(1.0)*4.0;
